refactor(base): split world_entry_appearance/dispatch/inventory-grant/player-load-core along handler/lifecycle seams
Tech-debt issue #529: split four oversized files in the base subsystem
(crate cimmeria-services) along natural seams. Behavior-preserving pure
code movement — re-export discipline keeps every external `use` path
identical and zero public-surface change.

world_entry_appearance.rs (1199) -> world_entry_appearance/:
- builders.rs    pure wire-arg builders + bundle composers + byte-exact tests
- client_ready.rs handle_on_client_ready + negative-logging/first-login tests
- cinematic.rs   send_cinematic / cancelMovie / appearance-spam guard
- mod.rs         re-exports

dispatch.rs (1174, codecov-ignored) -> dispatch/:
- chat.rs        sendPlayerCommunication / chatJoin / chatLeave / AFK / DND
- session.rs     logOff / cancelLogOff
- diagnostics.rs elementDataRequest / perfStats sinks
- tests.rs       moved integration suite (verbatim)
- mod.rs         constants + the dispatch match (delegates to handlers)
- codecov.yml ignore path widened dispatch.rs -> dispatch/** (precedent: connect_loop #244)

inventory/grant/mod.rs (704) -> slim mod.rs + siblings:
- validation.rs  normalize_item_ids + item_allows_container
- grant_item.rs  handle_grant_item

player_load/core.rs (701) -> core/mod.rs + submodules:
- player_data.rs    query_player_load_data
- inventory_items.rs query_inventory_items
- mod.rs            constants (CONTAINER_BANDOLIER, EQUIPMENT_CONTAINERS,
                    INVENTORY_ITEM_SELECT) + re-exports + the two test modules

No net-new tests: every moved function already had coverage. Test files
relocated verbatim with import-path fixes only. Doc anchors in
docs/testing/inventory/services.md repointed to the new file locations.

Part of #529.
